Initial commit adding first class support for DNS Update messages.
- Factor out common parts of DNS message processing into
  a base class.
- Make the Message and Update classes inherit from there
  and move various functions between classes
- New code in Update to support full update functionality.
- Changes in the way sections are handled in some circumstances
  as the section handling for some of the Update sections in
  particular is a little different.
